mongodb_archiver/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mongodb-order-archiver"
version = "2.0.0"
description = "Professional MongoDB order archiving system with Change Streams support"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.8"
keywords = ["mongodb", "archiving", "change-streams", "automation"]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Database",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "pymongo>=4.6.0",
    "python-dotenv>=1.0.0",
    "faker>=22.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/mongodb-order-archiver"
Documentation = "https://github.com/yourusername/mongodb-order-archiver/blob/main/README.md"
Repository = "https://github.com/yourusername/mongodb-order-archiver"

[project.scripts]
mongodb-archiver = "main:main"
mongodb-simulate = "simulate:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["."]
omit = ["tests/*", "venv/*", "demo.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
